# compress/flate: improve decompression speed