--- a/src/compress/flate/gen_inflate.go
+++ b/src/compress/flate/gen_inflate.go
@@ -56,7 +56,7 @@ func (f *decompressor) $FUNCNAME$() {
 
 	// Optimization. Compiler isn't smart enough to keep f.b, f.nb in registers,
 	// but is smart enough to keep local variables in registers, so use nb and b,
-	// inline call to moreBits and reassign b, nb back to f on return.
+	// dict reference and reassign b, nb back to f on return.
 	fnb, fb, dict := f.nb, f.b, &f.dict
 	switch f.stepState {
 	case stateInit:
@@ -125,6 +125,7 @@ readLiteral:
 			return
 		// otherwise, reference to older data
 		case v < 265:
+			// No extra bits 
 			length = v - (257 - 3)
 		case v < maxNumLit:
 			val := decCodeToLen[(v - 257)]
@@ -171,9 +172,6 @@ readLiteral:
 			// cases, the chunks slice will be 0 for the invalid sequence, leading it
 			// satisfy the n == 0 check below.
 			n := uint(f.hd.maxRead)
-			// Optimization. Compiler isn't smart enough to keep f.b,f.nb in registers,
-			// but is smart enough to keep local variables in registers, so use nb and b,
-			// inline call to moreBits and reassign b,nb back to f on return.
 			for {
 				for fnb < n {
 					c, err := fr.ReadByte()