@ianlancetaylor 2021-05-12
@volknanebo We don't use GitHub for code review.  If you want to make a comment, please make it at https://golang.org/cl/227737.  Thanks.