@klauspost 2022-09-07
Ping @ianlancetaylor - if there is interest for this in 1.20 it would be good to get started on CR.